Work around recent official header changes
Commit dcf7d4fb1e9ce5e219239443ca3f442c4c12c03d in the official OpenCL
headers repository (2024-06-18) defines `cl_version` unconditionally,
causing an issue when building clinfo due to the redefinition of the
symbol, even if with the same meaning.

A possible workaround (as suggested in the GitHub issue #92) is to move
our `CL_TARGET_OPENCL_VERSION` up to 300, and wrap our `cl_version`
definition around `CL_VERSION_3_0` being defined, but this causes
warnings when using any version of the unified headers before 3.0 was
finalized.

Instead, we leverage the fact that the commit that exposes `cl_version`
also exposes the `CL_MAKE_VERSION` macro. We still target 220, but we
fence the `cl_version` typedef with a check for the lack of a definition
for `CL_MAKE_VERSION`.

Tested with several version of the unified headers (the ancient
non-unified headers aren't a problem here), before and after the release
of 3.0, and before and after the 2024-06-18 change.